Fix composeMiddleware and move back into separate module
Discovered a subtle and confounding bug with composeMiddleware where
it only works if dispatch is the last argument. It did not combine
multiple middlewares into a single middlewares as advertised; it
only combined multiple middlewares with dispatch to create a new
dispatch. This didn't come up earlier because the codebase never
happened to use it in the former way.

This commit fixes the issue and adds a test for it.